pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mkclustering"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "tqdm>=4.66.0",
  "pandas>=2.0,<2.3",
  "numpy<2.0",
  "scikit-learn==1.3.2",
  "gensim==4.3.3",
  "conllu>=4.5.3",
  "requests==2.28.0",
  "matplotlib>=3.8.0",
  "PyYAML>=6.0.0",
  "fasttext-wheel>=0.9.2",
  "classla==2.2.1",
]

[project.optional-dependencies]
# install with: pip install -e ".[lemma]"
lemma = [
  "stanza>=1.8.2",
]

[project.scripts]
mkcli = "mkclustering.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
